190426 tilemap. pick & drag 구현 중

+++ Framework/Core/D3D11/BasicData/Config.h
@@ -57,6 +57,6 @@ struct TileData {
 	Vector2 TextureSize;
 	Vector2 SpriteOffset;
 	Vector2 SpriteSize;
-	float TilesetIndex;
+	int TilesetIndex;
 	float Padding;
 };
